+++ src/lib/Scenes/Home/Components/Sales/index.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { RefreshControl, SectionList, StyleSheet } from "react-native"
 import { createRefetchContainer, graphql, RelayRefetchProp } from "react-relay"
-import LotsByFollowedArtists from "./Components/LotsByFollowedArtists"
+// import LotsByFollowedArtists from "./Components/LotsByFollowedArtists"
 import { SaleList } from "./Components/SaleList"
 import { ZeroState } from "./Components/ZeroState"
 
@@ -69,11 +69,12 @@ class Sales extends React.Component<Props, State> {
         title: "Current Timed Auctions",
         renderItem: props => <SaleList {...props} />,
       },
-      {
-        data: [{ data: this.data.viewer }],
-        title: "Lots by Artists You Follow",
-        renderItem: props => <LotsByFollowedArtists title={props.section.title} viewer={props.item.data} />,
-      },
+      // TODO: https://github.com/artsy/collector-experience/issues/905
+      // {
+      //   data: [{ data: this.data.viewer }],
+      //   title: "Lots by Artists You Follow",
+      //   renderItem: props => <LotsByFollowedArtists title={props.section.title} viewer={props.item.data} />,
+      // },
     ]
 
     return (
@@ -97,7 +98,8 @@ export default createRefetchContainer(
         href
         live_start_at
       }
-      ...LotsByFollowedArtists_viewer
+      # TODO: https://github.com/artsy/collector-experience/issues/905
+      #...LotsByFollowedArtists_viewer
     }
   `,
   graphql`